feat(metric): 完善平台指标归一化与加权配置
- 支持使用小数 scoreWeight 计算平台加权平均分
- 测试配置改用 popularityMedian 和 popularityWeight
- 补充各平台热度归一化及聚合权重测试

+++ server/src/main/java/com/yukiani/server/config/PlatformConfig.java
@@ -14,7 +14,7 @@ public class PlatformConfig {
     private double popularityMedian = 10000;
 
     /** 归一化评分参与动画加权平均时使用的权重。 */
-    private int scoreWeight = 1;
+    private double scoreWeight = 1;
 
     /**
      * 评分平滑使用的归一化热度先验强度；未配置时不平滑。
